Lab4DB/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
Flasgger
dotenv
